added sign up page

+++ cms/users.php
@@ -6,7 +6,20 @@ include('includes/functions.php');
 secure();
 include('includes/header.php');
 
-if (isset($_GET['delete'])) {
+$currentUserId = $_SESSION['id'];
+
+$isAdmin = false;
+
+if ($stm = $connect->prepare('SELECT 1 from ADMIN where id = ?')) {
+    $stm->bind_param('i', $currentUserId);
+    $stm->execute();
+
+    $stm->store_result();
+    $isAdmin = $stm->num_rows > 0;
+    $stm->close();
+}
+
+if ($isAdmin && isset($_GET['delete'])) {
     if ($stm = $connect->prepare('DELETE FROM users WHERE id = ?')) {
         // $hashed = SHA1($_POST['password']);
         // $stm -> bind_param('ss', $_POST['email'], $hashed);
@@ -23,7 +36,17 @@ if (isset($_GET['delete'])) {
     }
 }
 
-if ($stm = $connect->prepare('SELECT * FROM users')) {
+if ($isAdmin) {
+    $sql = 'SELECT * FROM users';
+} else {
+    $sql = 'SELECT * FROM users WHERE id = ?';
+}
+
+if ($stm = $connect->prepare($sql)) {
+    if (!$isAdmin) {
+        $stm->bind_param('i', $currentUserId);
+    }
+
     $stm->execute();
     $result = $stm->get_result();
 
@@ -35,27 +58,31 @@ if ($stm = $connect->prepare('SELECT * FROM users')) {
                     <h1 class="h1-display-h1">User management</h1>
                     <table class="table table-striped table-hover">
                         <tr>
-                            <th>Id</th>
                             <th>Username</th>
                             <th>Email</th>
                             <th>Status</th>
-                            <th>Edit | Delete</th>
+                            <th>Edit</th>
+                            <th>Delete</th>
                         </tr>
 
                         <?php while ($record = mysqli_fetch_assoc($result)) { ?>
                             <tr>
-                                <td><?php echo $record['id']; ?></td>
                                 <td><?php echo $record['username']; ?></td>
                                 <td><?php echo $record['email']; ?></td>
                                 <td><?php echo $record['active']; ?></td>
                                 <td>
-                                    <a href="users_edit.php?id=<?php echo $record['id']; ?>">Edit</a>
-                                    <a href="users.php?delete=<?php echo $record['id']; ?>">Delete</a>
+                                    <a href="/cms/users_edit.php?id=<?php echo $record['id']; ?>">Edit</a>
+                                </td>
+                                <td>
+                                    <a href="/cms/users.php?delete=<?php echo $record['id']; ?>">Delete</a>
                                 </td>
                             </tr>
                         <?php } ?>
                     </table>
-                    <a href="users_add.php">Add new user</a>
+                    <?php if ($isAdmin) { ?>
+                        <a href="/cms/users_add.php">Add new user</a>
+
+                    <?php } ?>
                 </div>
             </div>
         </div>
